Hotfix/save transaction

+++ features/channels/src/main/java/info/alihabibi/channels/screens/ChannelsList.kt
@@ -47,7 +47,7 @@ fun ChannelsListDestination(
     ChannelsListScreen(
         channels = channels,
         onEditChannel = { channel ->
-            onEditChannel(channel.id)
+            onEditChannel(channel.id.toInt())
         },
         onAddNewChannel = {
             viewModel.onEvent(ChannelsUiIntent.ResetChannelsDraft)
